requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# AI/LLM Providers
anthropic==0.39.0
google-generativeai==0.8.3
openai==1.54.0

# Database & Persistence
# SQLite is built into Python - no additional dependency needed
pandas==2.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0
chardet==5.2.0

# Utilities
tiktoken==0.7.0
colorama==0.4.6
tqdm==4.66.1
jsonschema==4.21.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
